requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
beautifulsoup4==4.14.3
blis==1.3.3
catalogue==2.0.10
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cloudpathlib==0.23.0
confection==0.1.5
cryptography==46.0.5
cssselect==1.3.0
cymem==2.0.13
distro==1.9.0
exceptiongroup==1.3.1
feedfinder2==0.0.4
feedparser==6.0.12
filelock==3.19.1
google-api-core==2.29.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
googleapis-common-protos==1.72.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
jieba3k==0.35.1
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
lxml==6.0.2
lxml_html_clean==0.4.3
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
murmurhash==1.0.15
newspaper3k==0.2.8
nltk==3.9.2
numpy==2.0.2
oauthlib==3.3.1
openai==2.21.0
packaging==26.0
pandas==2.3.3
pillow==11.3.0
praw==7.8.1
prawcore==2.4.0
asyncpraw==7.7.0
preshed==3.0.12
proto-plus==1.27.1
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
requests-file==3.0.1
requests-oauthlib==2.0.0
rich==14.3.2
rsa==4.9.1
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smart_open==7.5.0
sniffio==1.3.1
soupsieve==2.8.3
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
textblob==0.19.0
thinc==8.3.9
tinysegmenter==0.3
tldextract==5.3.0
tqdm==4.67.3
tweepy==4.16.0
typer==0.23.1
typer-slim==0.23.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
update-checker==0.18.0
uritemplate==4.2.0
urllib3==2.6.3
wasabi==1.1.3
weasel==0.4.3
websocket-client==1.9.0
wrapt==2.1.1
yake==0.4.8
jellyfish==1.2.1
networkx==3.2.1
segtok==1.5.11
tabulate==0.9.0
redis==5.0.1
